TransactionFragment.kt ViewModel

+++ app/src/main/java/com/enigmacamp/goldmarket/data/model/CustomerBalance.kt
@@ -3,16 +3,16 @@ package com.enigmacamp.goldmarket.data.model
 import android.os.Parcel
 import android.os.Parcelable
 
-data class CustomerBalance(var customerId: String = "", var goldInGram: Int = 0) : Parcelable {
+data class CustomerBalance(var customerId: String = "", var goldInGram: Double = 0.0) : Parcelable {
     constructor(parcel: Parcel) : this(
         parcel.readString() ?: "",
-        parcel.readInt() ?: 0
+        parcel.readDouble() ?: 0.0
     ) {
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeString(customerId)
-        parcel.writeInt(goldInGram)
+        parcel.writeDouble(goldInGram)
     }
 
     override fun describeContents(): Int {
